requirements.txt
numpy>=1.24
pandas>=2.0
torch>=2.2
torchvision>=0.17
optuna>=3.5
matplotlib>=3.8
seaborn>=0.13
scikit-learn>=1.4
pyyaml>=6.0
tqdm>=4.66
